warning when debug does not contain prefiltered gyro data
I'd like to propose that the debug noise plots display a warning if the debug variable does not contain prefiltered gyro data. This would help avoid the confusion some people have when they have unknowingly selected the incorrect debug_mode in Betaflight, resulting in meaningless results in the debug subplots. An easy way to do this is to code when the 4th column of debug[3] is not empty indicating debug variable does not contain prefiltered gyro data. We could consider using debug_mode variable itself, but this wont work for certain earlier versions of BF. The solution I propose here involves 2 simple lines of code, the first (line 703) checks to see if debug[3] contains data (indicating incorrect debug_mode, since 4th column should be empty otherwise), and the second is a conditional statement (line 470) for presenting warning overlaid on debug subplots. There might be a more elegant way to do this, but this works.